Move map list generation logic
