+++ app/features/calendar/routes/calendar.new.tsx
@@ -36,7 +36,7 @@ import { useTranslation } from "~/hooks/useTranslation";
 import { useUser } from "~/modules/auth";
 import { requireUserId } from "~/modules/auth/user.server";
 import { i18next } from "~/modules/i18n";
-import { MapPool } from "~/modules/map-pool-serializer";
+import { MapPool } from "~/features/map-list-generator/core/map-pool";
 import { canEditCalendarEvent, canEnableTOTools } from "~/permissions";
 import calendarNewStyles from "~/styles/calendar-new.css";
 import mapsStyles from "~/styles/maps.css";
