+++ app/features/map-list-generator/core/map-pool-serializer/serializer.test.ts
