+++ app/features/map-list-generator/core/map-pool-serializer/types.ts
@@ -1,4 +1,4 @@
-import type { ModeShort, StageId } from "../in-game-lists";
+import type { ModeShort, StageId } from "../../../../modules/in-game-lists";
 
 export type MapPoolObject = Record<ModeShort, StageId[]>;
 export type ReadonlyMapPoolObject = Readonly<
