+++ app/features/sendouq/routes/q.looking.tsx
@@ -14,7 +14,7 @@ import { SubmitButton } from "~/components/SubmitButton";
 import { useIsMounted } from "~/hooks/useIsMounted";
 import { useTranslation } from "~/hooks/useTranslation";
 import { getUser, requireUser } from "~/modules/auth/user.server";
-import { MapPool } from "~/modules/map-pool-serializer";
+import { MapPool } from "~/features/map-list-generator/core/map-pool";
 import {
   parseRequestFormData,
   validate,
