+++ app/features/tournament/queries/upsertCounterpickMaps.server.ts
@@ -1,6 +1,6 @@
 import { sql } from "~/db/sql";
 import type { TournamentTeam } from "~/db/types";
-import type { MapPool } from "~/modules/map-pool-serializer";
+import type { MapPool } from "~/features/map-list-generator/core/map-pool";
 
 const deleteCounterpickMapsByTeamIdStm = sql.prepare(/* sql */ `
   delete from
