--- app/modules/map-list-generator/index.ts
@@ -1,3 +1,0 @@
-export { generateMapList } from "./map-list";
-export { modesOrder } from "./modes";
-export { mapPoolToNonEmptyModes, mapPoolListToMapPoolObject } from "./utils";
