--- app/modules/map-pool-serializer/README.md
@@ -1,3 +1,0 @@
-This module implements https://github.com/IPLSplatoon/IPLMapGen2/blob/splat3/url-serialization-docs.md
-
-It is a method to serialize/deserialize a map pool to be used with URL search params.
