--- app/modules/map-pool-serializer/index.ts
@@ -1,2 +1,0 @@
-export { MapPool } from "./map-pool";
-export type { MapPoolObject } from "./types";
